start finishing sign up process

+++ JoeysHomesApp/pages/Signup.jsx
@@ -16,20 +16,30 @@ export function loader({ request }) {
 
 export async function action({ request }) {
     const formData = await request.formData()
+
     const email = formData.get("email")
     const password = formData.get("password")
+    const confirmpassword = formData.get("confirmpassword")
+
+    if (password != confirmpassword) {
+        return "Passwords don't match, try again"
+    }
+
     const pathname = new URL(request.url)
         .searchParams.get("redirectTo") || "/"
     
     console.log("pathname (action)")
     console.log(pathname)
+
     try {
         const data = await signupUser({ email, password })
         console.log("data (Login client)")
         console.log(data)
         // TODO Do I need to get client info (email/password) returned here? Probably not
         // TODO Change 'loggedin' to fingerprint which ties to DB
+
         localStorage.setItem("loggedin", true)
+        
         // TODO If user is logged in, then upon subsequent requests to login page, either stay on current page or go to some account settings page
         return redirect(pathname)
     } catch(err) {
